pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-soc-analyst-env"
version = "0.1.0"
description = "SOC Analyst triage and investigation environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "requests>=2.25.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "soc_analyst_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["soc_analyst_env", "soc_analyst_env.server", "soc_analyst_env.scenarios"]

[tool.setuptools.package-dir]
"soc_analyst_env" = "."
"soc_analyst_env.server" = "server"
"soc_analyst_env.scenarios" = "scenarios"
